re-wrote config module, made progress on IPAM class
the config module has a much simpler and more robust structure now
thanks to ConfigParser and its .read(paths) method.

the IPAM class now has better flow and is better able to handle session
expiry and invalid credentials without have to raise an error and exit.
The idea here is to be able to have the IPAM class be able to behave
correctly even if it's used in a long-running script.
o Please enter the commit message for your changes. Lines starting