cli (almost) complete, progress made on deep search feature
The CLI script is mostly complete at this point. if called with a
subcommand, it runs that sub command and then exits. if run without a
subcommand, it drops into a shell that maintains a single IPAM instance
accross multiple subcommand calls. Need to do more testing to make sure
I'm not missing any edge cases, but overall I'm pretty happy with it.