deep search progresses
deep search has been implemented and works pretty well. two bugs need
to be addressed though:

in deep search, only entries where the start of the hostname matches
the keyword are selected. entries where the keyword is a partial in the
middle of the hostname string do not match

need to implement logic to remove duplicates from deep search results